weston: remove the definition of RECIPE_SYSROOT in .bbappend file
Issue: LIN10-3995

the process of building weston failed while try to do_configure. The log
shows that the required libraries are not in sysroot. While checking the
recipe-sysroot directory, we found it is empty though do_prepare_recipe_sysroot
success. So we checked the RECIPE_SYSROOT value and it is defined as
${STAGING_DIR}/${MACHINE} and the value is "build/tmp-glibc/sysroots/nxp-ls20xx".
This is not satisfied with the requirement of the yocto in wrlinux10. The solution
is simple: just remove the definition of RECIPE_SYSROOT and let yocto setup it.
The correct value should be something like below:
"build/tmp-glibc/work/aarch64-wrs-linux/weston/2.0.0-r0/recipe-sysroot".

(LOCAL REV: NOT UPSTREAM) -- Won't send it to meta-freescale.

Upstream-Status: Pending

Signed-off-by: Rui Wang <rui.wang@windriver.com>