u-boot-fslc-mxsboot-native: B!=S fix
commit 8e887cfcd810631680f5618b33eb0916cf995551 from
git://git.yoctoproject.org/meta-freescale

MACHINEs such as the imx23evk and imx28evk (i.e. the mxs MACHINEs)
failed to build because the recipe for u-boot-fslc-mxsboot-native wasn't
instrumented to support ${B} != ${S}, which is used as a result of inheriting
u-boot-fslc-common. In other words, the inherit was setting up ${B} != ${S},
but then u-boot-fslc-mxsboot didn't know what to do in this scenario.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>