u-boot-imx: resolve build issue with missing target
commit 6ca16f44b8a2d148e74cf9c1a2eb8790e0bb0d51 from
git://git.yoctoproject.org/meta-freescale

Since the replacement of u-boot-fw-utils with libubootenv in OE-Core,
U-Boot recipe requires an additional target to be built, which contains
the default environment.

This target has been introduced in U-Boot with version 2019.07 and is
missing in NXP fork of U-Boot source tree, therefore one patch from the
U-Boot source tree is backported in order to introduce a missing build
target.

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>