meta-freescale: introduce variable BSP_NXP_DERIVED and a BBFILES filter
Issue: LINUXEXEC-15236

Introduce a filter function based on MACHINE and a variable
BSP_NXP_DERIVED which can be set from another layer to append
to the filter function. So that the customer can create a new
additional layer with their new machine configuration by appending
bsp name to the variable in the local.conf. And the filter function
make the *.bb and *.bbappend of meta-freescale valid only for bsps
in BSP_NXP_DERIVED.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>